+++ src/presentation/components/PetitionList/VPetitionList.tsx
@@ -70,6 +70,7 @@ const statusColor = {
   '답변 대기중': '#DF3127',
   '답변 완료': '#008DD5',
   '승인 반려': '#000000',
+  '청원 만료': '#000000',
 };
 
 const VPetitionList = ({ isLoading, petitions, type }: vPetitionListProps): JSX.Element => {
@@ -86,8 +87,9 @@ const VPetitionList = ({ isLoading, petitions, type }: vPetitionListProps): JSX.
         </HeaderItem>
         <StLine />
         {petitions.map((petition) => {
-          const { released, agreeCount, answered, rejected, id, categoryName, title, createdAt, tempUrl } = petition;
-          const status = checkPetitionStatus(released, agreeCount, answered, rejected);
+          const { released, agreeCount, answered, rejected, id, categoryName, title, createdAt, tempUrl, expired } =
+            petition;
+          const status = checkPetitionStatus(released, agreeCount, answered, rejected, expired);
           return (
             <div key={'petition_item_' + id}>
               <PetitionItem>
