+++ src/presentation/pages/AnswerPetition.tsx
@@ -1,4 +1,4 @@
-import { getPetitionById } from '@api/petitionAPI';
+import { getPetitionById } from '@api/petitionQueryAPI';
 import { Title, Wrapper } from '@components/common';
 import VPetition from '@components/common/VPetition';
 import WriteAnswer from '@components/WriteAnswer';
